requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
pyyaml>=6.0.0
tushare>=1.2.89
python-telegram-bot>=21.0
fastapi>=0.110.0
uvicorn>=0.29.0
